+++ example/lib/main.dart
@@ -3,5 +3,5 @@ import 'package:flutter/material.dart';
 import 'src/app.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }