+++ example/lib/src/activities/loading_activity.dart
@@ -1,36 +1,50 @@
-
 import 'package:audio_in_app/audio_in_app.dart';
 import 'package:flutter/material.dart';
 
 class LoadingActivity extends StatefulWidget {
+  const LoadingActivity({super.key});
+
   @override
   State<LoadingActivity> createState() => _LoadingActivityState();
 }
 
 class _LoadingActivityState extends State<LoadingActivity> {
-  AudioInApp _audioInApp = AudioInApp();
+  final AudioInApp _audioInApp = AudioInApp();
 
   @override
   void initState() {
     super.initState();
-    Future.delayed(Duration(milliseconds: 1500)).then((value) =>   goToMain());
+    Future.delayed(const Duration(milliseconds: 1500)).then((_) => _goToMain());
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       body: Center(
-        child: Text('Loading Activity'),
+        child: Text('Loading audio...'),
       ),
     );
   }
 
+  Future<void> _goToMain() async {
+    await _audioInApp.createNewAudioCache(
+      playerId: 'button',
+      route: 'audio/button.wav',
+      audioInAppType: AudioInAppType.determined,
+    );
+    await _audioInApp.createNewAudioCache(
+      playerId: 'intro1',
+      route: 'audio/intro_1.wav',
+      audioInAppType: AudioInAppType.background,
+    );
+    await _audioInApp.createNewAudioCache(
+      playerId: 'intro2',
+      route: 'audio/intro_2.wav',
+      audioInAppType: AudioInAppType.background,
+    );
 
-  Future<void> goToMain() async {
-    await _audioInApp.createNewAudioCache(playerId: 'button', route: 'audio/button.wav', audioInAppType: AudioInAppType.determined);
-    await _audioInApp.createNewAudioCache(playerId: 'intro1', route: 'audio/intro_1.wav', audioInAppType: AudioInAppType.background);
-    await _audioInApp.createNewAudioCache(playerId: 'intro2', route: 'audio/intro_2.wav', audioInAppType: AudioInAppType.background);
-
-    Navigator.pushReplacementNamed(context, 'main');
+    if (mounted) {
+      Navigator.pushReplacementNamed(context, 'main');
+    }
   }
 }