+++ example/lib/src/app.dart
@@ -1,23 +1,16 @@
-
 import 'package:example/src/routes/routes.dart';
 import 'package:flutter/material.dart';
 
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
 
-class MyApp extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() {
-    return _MyApp();
-  }
-}
-
-class _MyApp extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Example App',
+      title: 'audio_in_app Example',
       debugShowCheckedModeBanner: false,
       initialRoute: 'loading',
-      routes: getAplicationRouter(),
+      routes: getApplicationRouter(),
     );
   }
 }