+++ example/lib/src/routes/routes.dart
@@ -2,9 +2,9 @@ import 'package:example/src/activities/loading_activity.dart';
 import 'package:example/src/activities/main_activity.dart';
 import 'package:flutter/material.dart';
 
-Map<String, WidgetBuilder> getAplicationRouter(){
+Map<String, WidgetBuilder> getApplicationRouter() {
   return <String, WidgetBuilder>{
-    'loading'                     : (BuildContext context) => LoadingActivity(),
-    'main'                        : (BuildContext context) => MainActivity(),
+    'loading': (BuildContext context) => const LoadingActivity(),
+    'main': (BuildContext context) => const MainActivity(),
   };
 }